pyproject.toml
[project]
name = "task"
version = "0.1.0"
description = "Agentic entity discovery system"
readme = "README.md"
requires-python = ">=3.13"
dependencies = [
    "atomic-agents>=2.7.4",
    "ddgs>=9.0.0",
    "google-genai>=1.39.1",
    "instructor>=1.8.0",
    "openai>=2.30.0",
    "python-dotenv>=1.1.0",
    "requests>=2.32.0",
    "trafilatura>=2.0.0",
]
